test(state): fix RxStateService spec. move all specs to spec folder
