+++ libs/state/spec/rx-state.service.spec.ts
@@ -1,3 +1,4 @@
+import { Injector, runInInjectionContext } from '@angular/core';
 import { fakeAsync, TestBed } from '@angular/core/testing';
 import { RxState } from '../src/lib/rx-state.service';
 import { select } from '@rx-angular/state/selections';
@@ -13,9 +14,9 @@ import { map, pluck, switchMap, take, takeUntil } from 'rxjs/operators';
 import { TestScheduler } from 'rxjs/testing';
 import { createStateChecker } from './fixtures';
 
-function setupState<T extends object>(cfg: { initialState?: T }) {
+function setupState<T extends object>(cfg: { initialState?: T } = {}) {
   const { initialState } = { ...cfg };
-  const state = new RxState<T>();
+  const state = TestBed.inject(RxState);
   if (initialState) {
     state.set(initialState);
   }
@@ -38,12 +39,15 @@ beforeEach(() => {
 
 describe('RxStateService', () => {
   let service: RxState<PrimitiveState>;
+  let injector: Injector;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       teardown: { destroyAfterEach: true },
+      providers: [RxState],
     });
-    service = setupState({});
+    service = setupState();
+    injector = TestBed.inject(Injector);
   });
 
   it('should be created', () => {
@@ -62,7 +66,7 @@ describe('RxStateService', () => {
 
   describe('State', () => {
     it('should create new instance', () => {
-      const state = new RxState<PrimitiveState>();
+      const state = setupState();
       expect(state).toBeDefined();
     });
   });
@@ -70,24 +74,24 @@ describe('RxStateService', () => {
   describe('$', () => {
     it('should return NO empty state after init when subscribing late', () => {
       testScheduler.run(({ expectObservable }) => {
-        const state = setupState({});
-        expectObservable(state.$).toBe('');
+        expectObservable(service.$).toBe('');
       });
     });
 
     it('should return No changes when subscribing late', () => {
       testScheduler.run(({ expectObservable }) => {
-        const state = new RxState<PrimitiveState>();
-        state.subscribe();
+        runInInjectionContext(injector, () => {
+          const state = service;
+          state.subscribe();
 
-        state.set({ num: 42 });
-        expectObservable(state.$.pipe(pluck('num'))).toBe('');
+          state.set({ num: 42 });
+          expectObservable(state.$.pipe(pluck('num'))).toBe('');
+        });
       });
     });
 
     it('should return new changes', () => {
-      const state = new RxState<PrimitiveState>();
-      state.subscribe();
+      const state = service;
       state.set({ num: 42 });
       const slice$ = state.$.pipe(select('num'));
       let i = -1;
@@ -100,31 +104,34 @@ describe('RxStateService', () => {
   describe('stateful with select', () => {
     it('should return empty state after init when subscribing late', () => {
       testScheduler.run(({ expectObservable }) => {
-        const state = setupState({});
-        expectObservable(state.select()).toBe('');
+        expectObservable(service.select()).toBe('');
       });
     });
 
     it('should return changes when subscribing late', () => {
       testScheduler.run(({ expectObservable }) => {
-        const state = new RxState<PrimitiveState>();
-        state.subscribe();
+        runInInjectionContext(injector, () => {
+          const state = new RxState<PrimitiveState>();
+          state.subscribe();
 
-        state.set({ num: 42 });
-        expectObservable(state.select('num')).toBe('n', { n: 42 });
+          state.set({ num: 42 });
+          expectObservable(state.select('num')).toBe('n', { n: 42 });
+        });
       });
     });
 
     it('should return new changes', () => {
       testScheduler.run(({ expectObservable }) => {
-        const state = new RxState<PrimitiveState>();
-        state.subscribe();
-        state.set({ num: 42 });
-        const slice$ = state.select('num');
-        let i = -1;
-        const valuesInOrder = [42, 777];
-        slice$.subscribe((next) => expect(next).toBe(valuesInOrder[++i]));
-        state.set({ num: 777 });
+        runInInjectionContext(injector, () => {
+          const state = new RxState<PrimitiveState>();
+          state.subscribe();
+          state.set({ num: 42 });
+          const slice$ = state.select('num');
+          let i = -1;
+          const valuesInOrder = [42, 777];
+          slice$.subscribe((next) => expect(next).toBe(valuesInOrder[++i]));
+          state.set({ num: 777 });
+        });
       });
     });
   });
@@ -190,11 +197,13 @@ describe('RxStateService', () => {
 
       it('should return initial state', () => {
         testScheduler.run(({ expectObservable }) => {
-          const state = new RxState<PrimitiveState>();
-          state.subscribe();
+          runInInjectionContext(injector, () => {
+            const state = new RxState<PrimitiveState>();
+            state.subscribe();
 
-          state.set({ num: 42 });
-          expectObservable(state.select('num')).toBe('s', { s: 42 });
+            state.set({ num: 42 });
+            expectObservable(state.select('num')).toBe('s', { s: 42 });
+          });
         });
       });
     });
